Temporarily comment out parse_dsn() in the bounce parser.
Add "save_denial_checksums" configuration variable.  This
determines whether or not message body checksums of messages
denied by the access rules will be saved in the duplicate
database.  It is off by default.

Add "reason" action to access_rules.  This allows additional
information to be added to reply files for denied and stalled
requests.

Cause "unset" access rule action to set a variable to zero
instead of deleting it.  This is used by the denial checksum
code to determine if the checksum variables were reset
by the access rules or never defined in the first place.
Thus, no checksum-saving side effects are caused by changing
the checksum variables.

Include the checksums in the access "avars" data used by
Mj::Resend::post.

Cause a 450 return code from RCPT TO to defer delivery for
that recipient.