(new) Treat the registry as a subscriber list for the GLOBAL list. This allows the Deliver module to distribute messages to addresses in the registry.
(announce) New command.  Distribute a file to one
or more classes of subscribers.  Each message is
sent as a probe, with the recipient in the To: header.
This could be used for periodic reminders, or
announcements that need to reach 'digest' and 'nomail'
subscribers immediately.

(archive_chunk) Allow the type of digest for the
archive-get command to be specified in the mode ("-mime"
for mime; text otherwise.)

(report) New, iterated command.  Obtain raw or summarized
data for recent log entries for a mailing list.
The data can be constrained by action or by date.