archive_url:  New configuration variable, used by mj_wwwusr and the message_headers setting (ARCURL substitution in Mj::Resend::_post.)
sublists:  New configuration variable.
The lists-aux command will only display auxiliary lists
explicity mentioned in this setting.  The createlist-regen
command will only create mail aliases for sublists included
in this setting.  Each line of the array consists of
a sublist name, followed by an optional description.
The description appears in the output of the lists-aux command.